# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Biometric Core Backend API
  version: 0.1.0
  description: >-
    REST API for biometric enrollment, authentication, and device management.
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    XRequestId:
      description: Correlation ID echoed back on all responses.
      schema: { type: string }
    IdempotencyReplayed:
      description: True when a request with the same Idempotency-Key was replayed and a cached result is returned.
      schema: { type: boolean }
...
  parameters:
    XRequestId:
      in: header
      name: X-Request-Id
      required: false
      schema: { type: string }
      description: Optional request correlation ID. Echoed back in responses.
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string }
      description: Idempotency key for safely retrying mutating requests.
    Cursor:
      in: query
      name: cursor
      required: false
      schema: { type: string }
      description: Opaque pagination cursor.
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 250
        default: 25
      description: Page size (default 25, max 250).
...
  responses:
    Problem:
      description: Error response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    Meta:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        limit: { type: integer, minimum: 1, maximum: 250 }
    ProblemDetails:
      type: object
      required: [type, title, status, traceId]
      properties:
        type: { type: string, description: A URI reference that identifies the problem type }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Machine-readable error code }
        traceId: { type: string, description: Correlation/trace identifier }
...
    Device:
      type: object
      required: [id, userId, credentialId, active, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        credentialId: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    EnrollmentChallengeInput:
      type: object
      description: >
        Optional parameters for creating an enrollment challenge. The user to enroll is derived
        from the authenticated access token; clients may provide an optional device label.
      properties:
        deviceName:
          type: string
          maxLength: 128
          description: Optional human-friendly label for the device being enrolled.
    ChallengeResponse:
      type: object
      required: [challengeId, publicKeyCredentialOptions]
      properties:
        challengeId: { type: string }
        publicKeyCredentialOptions:
          type: object
          description: WebAuthn options as per spec (opaque to clients).
          additionalProperties: true
...
    EnrollmentVerifyInput:
      type: object
      required: [challengeId, credential]
      properties:
        challengeId: { type: string }
        credential:
          type: object
          description: WebAuthn attestation response from the client.
          additionalProperties: true
    EnrollmentVerifyResponse:
      type: object
      required: [credentialId, deviceId]
      properties:
        credentialId: { type: string }
        deviceId: { type: string }
...
    EnvelopeDevices:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Device' }
        meta: { $ref: '#/components/schemas/Meta' }
    EnvelopeChallengeResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/ChallengeResponse' }
...
    EnvelopeEnrollmentVerifyResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/EnrollmentVerifyResponse' }
paths:
...
  /v1/enroll/challenge:
    post:
      tags: [Enrollment]
      summary: Create enrollment challenge (WebAuthn attestation)
      operationId: enrollCreateChallenge
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollmentChallengeInput' }
      responses:
        '200':
          description: Challenge issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvelopeChallengeResponse' }
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
...
  /v1/enroll/verify:
    post:
      tags: [Enrollment]
      summary: Verify enrollment attestation and store credential
      operationId: enrollVerify
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollmentVerifyInput' }
      responses:
        '200':
          description: Device enrolled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvelopeEnrollmentVerifyResponse' }
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/devices:
    get:
      tags: [Devices]
      summary: List devices for current user
      operationId: listDevices
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvelopeDevices' }
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /v1/devices/{id}:
    delete:
      tags: [Devices]
      summary: Revoke a device by id
      operationId: revokeDevice
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
...